backend/requirements.txt
alembic==1.11.1 
aniso8601==9.0.1
attrs==23.1.0 
blinker==1.6.2 
click==8.1.5 
flask==2.3.2
flask-cors==4.0.0
flask-migrate==4.0.4
flask-restx==1.1.0
flask-sqlalchemy==3.0.5
flask-swagger-ui==4.11.1
iniconfig==2.0.0 
itsdangerous==2.1.2 
jinja2==3.1.2 
jsonschema==4.18.3 
jsonschema-specifications==2023.6.1 
mako==1.2.4 
markupsafe==2.1.3 
packaging==23.1 
pluggy==1.2.0 
psycopg2-binary==2.9.6
pytest==7.4.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
referencing==0.29.1 
rpds-py==0.8.11 
sqlalchemy==2.0.19 
typing-extensions==4.7.1 
werkzeug==2.3.6 
